Keep channel reference sharing on the existing Postgres adapter
